% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoPriorCVDBleedRisk.R
\name{MajorBleedRisk}
\alias{MajorBleedRisk}
\title{PREDICT Major Bleed (2018) Risk Score for People Without Prior CVD}
\usage{
NoPriorCVDBleedRisk(dat, sex, age, eth, smoker, nzdep, af, familyhx, diabetes,
                       sbp, tchdl, lld, bpl, cancer, gibleed, puddiag, alcohol,
                       liver, puddrug, nsaid, steroids, ssri,...)
}
\arguments{
\item{dat}{A data.frame or data.table containing input data. Optional. See Details.}

\item{sex}{Sex or gender - input as labels M, Male, F, Female; or encode binary where 1 is male and 0 is female}

\item{age}{Age - input as numeric value between 30 and 74. See age details if outside of this range.}

\item{eth}{Ethnicity - input as label or encoded value. See ethnicity details for all possible inputs.}

\item{smoker}{Smoking status - input as labels (or encode as) "Ex smoker" (2), "Ex-smoker" (2), "Ex" (2), "Current Smoker" (1), "Current" (1), "Smoker" (1), "Y" (1), "Yes" (1)}

\item{nzdep}{Index of socioeconomic deprivation, specifically the New Zealand Deprivation Index - input as numeric quintile value between 1 (least deprived) and 5 (most deprived).}

\item{familyhx}{Family history of premature CVD - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{diabetes}{Diabetes status - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{sbp}{Systolic blood pressure - input as numeric value representing measured systolic blood pressure in mmHg}

\item{tchdl}{Total-HDL cholesterol ratio - input as numeric value representing most recent value of total:HDL cholesterol}

\item{lld}{Receiving lipid lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{bpl}{Receiving at least one blood pressure lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{cancer}{Any prior primary malignancy excluding squamous and basal cell skin cancers}

\item{gibleed}{Any prior gastrointestinal bleeding - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{puddiag}{Any prior non-bleeding and non-perforated peptic ulcer disease - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{alcohol}{Chronic high use of alcohol - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{liver}{Chronic liver disease or pancreatitis - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{puddrug}{Receiving at peptic ulcer disease medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{nsaid}{Receiving at least one nonsteroidal anti-inflammatary drug - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{steroids}{Receiving at least one corticosteroid medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{ssri}{Receiving at least one selective serotonine reuptake inhibitor - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{...}{Set decimal place for integers. Default is 4. Optional.}
}
\description{
\code{NoPriorCVDBleedRisk} calculates the 5 year risk of major bleeding (gastrointestinal, intracranial, and other bleeds), for people without cardiovascular disease.
If a dataset of input values are not supplied, then individual values for each coefficient can be specified. If a dataset of input values are supplied, then a risk
estimate is produced for each row of data, resulting in a numeric vector of the same length. A specific format is required for each input value. Encoding may be
required. See arguments.
}
\details{
When the parameter \code{dat} is supplied using a dataset, then parameters take variable names as input. For example, when a dataset is supplied, the parameter \code{age} requires the variable name \code{index_age} as input from the dataset.
When the parameter \code{dat} is not supplied, then parameters take actual values or labels as input. For example, when \code{dat} is not supplied, the parameter \code{age} requires a single numeric value between 30 and 79. This method calculates the 5-year risk estimate for a single individual.
}
\section{Age}{

The risk prediction equations were developed from a cohort of people aged 30 to 79 years. People aged 18-29 years or 80 years and older, are outside the range used
to derive the equation and so risk will be even more of an approximation. To be consistent with equations for primary prevention in this suite of scores, the
function will calculate ages 18-29 as 30; and ages 80-110 as 80.
}

\section{Ethnicity}{

The co-efficients for ethnicity apply only to the following groups: European, Maori, Pacific, Indian, and Asian. Individuals with ethnicity labels (or codes) that fall outside of these categories will not recieve a risk estimate.
To obtain a risk estimate, ensure that the ethnicity parameter is either labelled (not case-sensitive) or encoded as one of the following:
\itemize{
\item NZ European, European, NZEO, Euro, E, 1, 10, 11, 12
\item Maori, NZMaori, NZ Maori, M, 2, 21
\item Pacific, Pacific Islander, PI, P, 3, 30, 31, 32, 33, 34, 35, 36, 37
\item Indian, Fijian Indian, South Asian, IN, I, 43
\item Asian, Other Asian, SE Asian, East Asian, Chinese, ASN, A, 4, 40, 41, 42, 44
}
}

\section{Value}{

Returns either a single bleeding risk estimate or a numeric vector of bleeding risk estimates.
}

\examples{
# As calculator (dataset not provided)
NoPriorCVDBleedRisk(sex=0, age=55, eth=21, smoker=1, nzdep=5, af=0, familyhx=1, diabetes=1, sbp=130, tchdl=5, lld=1,
                    bpl=1, cancer=1, gibleed=1,puddiag=1, alcohol=0, liver=0, puddrug=0, nsaid=1, steroids=1, ssri=0)

# As vectoriser (dataset provided)
NoPriorCVDBleedRisk(dat=DT, sex=sex, age=index_age, eth=eth_vars, smoker=smoking_status, nzdep=nzdep, af=af, familyhx=family_hx, diabetes=dm,
                    sbp=index_sbp, tchdl=tchdl, lld=lld, bpl=bpl, cancer=hx_cancer, gibleed=hx_gibleed, puddiag=hx_hud, alcohol=alc, liver=hx_liver,
                    puddrug=pudmx, nsaid=nsaid, steroids=steroidmx, ssri=ssri)

}
\references{
Selak V, Jackson R, Poppe K, et al. Predicting Bleeding Risk to Guide Aspirin Use for the Primary Prevention of Cardiovascular Disease: A Cohort Study. Ann Intern Med. 2019;170:357–368. doi: https://doi.org/10.7326/M18-2808

Full Article: \url{https://www.annals.org/aim/fullarticle/doi/10.7326/M18-2808}

Using the bleeding risk equation developed by Selak et al (2019), I developed a web-based risk calculator that provides clinicians with an individualised estimate of the CVD benefit and bleeding harms of aspirin for their patients without established CVD.

For the calculator: \url{https://aspirinbenefitharmcalculator.shinyapps.io/calculator/}
}
\seealso{
\code{\link{NoPriorCVDRisk}} Creates a 5 year CVD risk estimate for people without prior CVD using the published Lancet equation.

\code{\link{NoPriorCVDRisk_BMI}} Creates a 5 year CVD risk estimate for people without prior CVD using the Ministry of Health's HISO equation containing BMI.

\code{\link{PriorT2DRisk}} Creates a 5 year CVD risk estimate for people with prior Type-II diabetes using the Ministry of Health's HISO equation.

\code{\link{PriorCVDRisk}} Creates a 5 year CVD risk estimate for people with prior CVD using the published Heart equation.

\code{\link{PolicyCVDRisk}} Creates a 5 year CVD policy risk estimate for people in the general population using the publish IJE equation.

\code{\link{PostACSRisk}} Creates a 5 year CVD risk estimate for people after an ACS event using the published Heart equation.
}
\author{
Billy Wu (R developer) and Vanessa Selak (Principal Investigator)
}
