% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamCheck.R
\name{ParamCheck}
\alias{ParamCheck}
\title{General Parameter Checking for PredictRiskScores}
\usage{
ParamCheck(params, ignore.age)
}
\arguments{
\item{input}{list of all input values inherited from the parent.frame}

\item{vars}{character vector of required parameters from the parent.frame}

\item{call}{name of the function being called in the parent.frame}

\item{is.table}{logical; whether a dataset is provided in the parent.frame}

\item{allow.age}{logical; whether ages of between 18-29 and 75-110 are calculated}

\item{allow.na}{logical; whether \code{na} is allowed for binary variables including smoking. If TRUE, then \code{na} is treated as 0.}
}
\value{
\code{ParamCheck} returns warning messages if any invalid values are detected. In cases where input data is invalid, \code{ParamCheck} will
stop the function's operation and return an error message.
}
\description{
\code{ParamCheck} evaluates input parameters for names, values, and availability to ensure that a risk calculation can be carried out; else provide warnings, else stopped.
}
\details{
\code{ParamCheck} identifies the following errors and invalid inputs:
\enumerate{
\item missing arguments - operations stops
\item missing input data - operation stops
\item invalid sex, age, ethnicity, nzdep inputs
\item invalid binary inputs
\item invalid numeric inputs
\item invalid time inputs
}
}
\author{
Billy Wu (R Developer)
}
