% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostACSRisk.R
\name{PostACSRisk}
\alias{PostACSRisk}
\title{PREDICT CVD (2018) Risk Score for People After an ACS Event}
\usage{
PostACSRisk(dat, sex, age, eth, nzdep, smoker, diabetes,
           af, hf, acsdays, acstype, bmi, sbp, tchdl, hba1c,
           scr, bpl, lld, athromb,...)
}
\arguments{
\item{dat}{A data.frame or data.table containing input data. Optional. See Details.}

\item{sex}{Sex or gender - input as labels M, Male, F, Female; or encode binary where 1 is male and 0 is female}

\item{age}{Age - input as numeric value between 30 and 79. See age details if outside of this range.}

\item{eth}{Ethnicity - input as label or encoded value. See ethnicity details for all possible inputs.}

\item{nzdep}{Index of socioeconomic deprivation, specifically the New Zealand Deprivation Index - input as numeric quintile value between 1 (least deprived) and 5 (most deprived)}

\item{smoker}{Current smoker - input as labels "Y", "Yes", "Smoker", or encode binary where 1 is "Yes"}

\item{diabetes}{Diabetes status - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{af}{Atrial fibrillation status - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{hf}{Heart failure - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{acsdays}{Time since last the most recent ACS event - input as numeric value representing days. If the event was angina or peripheral vascular disease, then enter 1826. If the date of most recent CVD event was not known, then keep as NA.}

\item{acstype}{Prior ACS was NSTEMI or STEMI - input as labels (or encode as) NSTEMI (1), NONSTEMI (1) or STEMI(2)}

\item{bmi}{Body mass index - input as numeric value representing BMI in kg/m^2}

\item{sbp}{Systolic blood pressure - input as numeric value representing measured systolic blood pressure in mmHg}

\item{tchdl}{Total-HDL cholesterol ratio - input as numeric value representing most recent value of total:HDL cholesterol}

\item{hba1c}{HbA1C - input as numeric value representing most recent value of HbA1c in mmol/mol}

\item{scr}{Serum creatinine - input as numeric value representing most recent value of serum creatinine in micromol/L}

\item{bpl}{Receiving at least one blood pressure lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{lld}{Receiving lipid lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{athromb}{Receiving antiplatelet or anticoagulant medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{...}{Set decimal place for integers. Default is 4. Optional.}
}
\description{
\code{PostACSRisk} calculates the 5 year risk of cardiovascular disease (CVD) (hospitalisation for acute coronary syndrome, heart failure, stroke or other cerebrovascular disease, peripheral vascular death, cardiovascular death),
for people with a history of atherosclerotic CVD. If a dataset of input values are not supplied, then individual values for each coefficient can be specified. If a dataset of input values are supplied, then a risk estimate is produced for each row of data, resulting in a numeric vector of the same length.
A specific format is required for each variable input value. Encoding may be required. See arguments.
}
\details{
When the parameter \code{dat} is supplied using a dataset, then parameters take variable names as input. For example, when a dataset is supplied, the parameter \code{age} requires the variable name \code{index_age} as input from the dataset.
When the parameter \code{dat} is not supplied, then parameters take actual values or labels as input. For example, when \code{dat} is not supplied, the parameter \code{age} requires a single numeric value between 30 and 79. This method calculates the 5-year risk estimate for a single individual.
}
\section{Age}{

The primary prevention risk prediction equations were developed from a cohort of people aged 30 to 74 years who were eligible for CVD risk prediction according to the 2003 CVD risk assessment and management guidelines and subsequent updates (New Zealand Guidelines Group 2003).
People aged 18-29 years and 80 years and older, the equation will only provide a very approximate estimate. However, a risk calculation may be potentially useful to guide clinical decision making.
As such, the equation will calculate ages 18-29 as 30; and ages 80-110 as 80.
}

\section{Ethnicity}{

The co-efficients for ethnicity apply only to the following groups: European, Maori, Pacific, Indian, and Asian. Individuals with ethnicity labels (or codes) that fall outside of these categories will not recieve a risk estimate.
To obtain a risk estimate, ensure that the ethnicity parameter is either labelled (not case-sensitive) or encoded as one of the following:
\itemize{
\item NZ European, European, NZEO, Euro, E, 1, 10, 11, 12
\item Maori, NZMaori, NZ Maori, M, 2, 21
\item Pacific, Pacific Islander, PI, P, 3, 30, 31, 32, 33, 34, 35, 36, 37
\item Indian, Fijian Indian, South Asian, IN, I, 43
\item Asian, Other Asian, SE Asian, East Asian, Chinese, ASN, A, 4, 40, 41, 42, 44
}
}

\section{Value}{

Returns either a single CVD risk estimate or a numeric vector of CVD risk estimates.
}

\examples{
# As a calculator (dataset not provide)
PostACSRisk(sex="F", age=65, eth="Indian", nzdep=5, smoker=0, diabetes=0,
            af=0, hf=1, acsdays=65, acstype="NSTEMI", bmi=NA, sbp=118,
            tchdl=3.3, hba1c=NA, scr=52, bpl=1, lld=1, athromb=1)

# As Vectoriser (dataset provided)
PostACSRisk(TEST, sex=sex, age=age, eth=eth, nzdep=nzdep, smoker=smoker, diabetes=diabetes, af=af, hf=hf, acsdays=days, acstype=acs_type,
             bmi=bmi, sbp=sbp, tchdl=tchdl, hba1c=hba1c, scr=scr, bpl=bpl, lld=lld, athromb=athromb)


}
\references{
Poppe KK, Doughty RN, Wells S, et al. Development and validation of a cardiovascular risk score for patients in the community after acute coronary syndromeHeart Published Online First: 10 December 2019. doi: 10.1136/heartjnl-2019-315809

Full Article: \url{https://heart.bmj.com/content/early/2019/12/10/heartjnl-2019-315809.full}

Toll Free: \url{https://heart.bmj.com/content/heartjnl/early/2019/12/10/heartjnl-2019-315809.full.pdf?ijkey=B9NMccWMr793Ixj&keytype=ref}
}
\seealso{
\code{\link{NoPriorCVDRisk}} Creates a 5 year CVD risk estimate for people without prior CVD using the published Lancet equation.

\code{\link{NoPriorCVDRisk_BMI}} Creates a 5 year CVD risk estimate for people without prior CVD using the Ministry of Health's HISO equation containing BMI.

\code{\link{PriorT2DRisk}} Creates a 5 year CVD risk estimate for people with prior Type-II diabetes using the Ministry of Health's HISO equation.

\code{\link{MajorBleedRisk}} Creates a 5 year major bleeding risk estimate for people without prior CVD using the published AnnIntMed equation.

\code{\link{PriorCVDRisk}} Creates a 5 year CVD risk estimate for people with prior CVD using the published Heart equation.

\code{\link{PolicyCVDRisk}} Creates a 5 year CVD policy risk estimate for people in the general population using the publish IJE equation.
}
\author{
Billy Wu (R developer) and Katrina Poppe (Principle Investigator)
}
