% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorT2DRisk.R
\name{PriorT2DRisk}
\alias{PriorT2DRisk}
\title{PREDICT CVD Type-II Diabetes (2018.1) Risk Score for People Without Prior CVD}
\usage{
PriorT2DRisk(dat, sex, age, eth, nzdep, smoker, af, familyhx, sbp, tchdl, bmi,
           years, egfr, acr, hba1c, oral, insulin, bpl, lld, athromb,...)
}
\arguments{
\item{dat}{A data.frame or data.table containing input data. Optional. See Details.}

\item{sex}{Sex or gender - input as labels M, Male, F, Female; or encode binary where 1 is male and 0 is female}

\item{age}{Age - input as numeric value between 30 and 74}

\item{eth}{Ethnicity - input as labels (or encode as) "European" (1), "Maori" (2), "Pacific" (3), "Chinese" (42), "Indian" (43), "Fijian Indian" (43), "Other South Asian" (43), or "Other Asian" (4).}

\item{nzdep}{Index of socioeconomic deprivation, specifically the New Zealand Deprivation Index - input as numeric quintile value between 1 (least deprived) and 5 (most deprived).}

\item{smoker}{Current smoker - input as labels "Y", "Yes", "Smoker", or encode binary where 1 is "Yes"}

\item{af}{Atrial fibrillation status - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{familyhx}{Family history of premature CVD - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{sbp}{Systolic blood pressure - input as numeric value representing measured systolic blood pressure in mmHg}

\item{tchdl}{Total-HDL cholesterol ratio - input as numeric value representing most recent value of total:HDL cholesterol}

\item{bmi}{Body mass index - input as numeric value representing BMI in kg/m^2}

\item{years}{Years since diagnosis of type 2 diabetes - input as numeric value representing years}

\item{egfr}{Calculated or measured estimated glomerular filtration rate - input as numeric value representing most recent value eGFR value in mL/min/1.73m2}

\item{acr}{Calculated or Measured albumin to creatinine ratio - input as numeric value representing most recent ACR value in mg/mmol}

\item{hba1c}{Haemoglobin A1C - input as numeric value representing most recent value of HbA1c in mmol/mol}

\item{oral}{On oral hypoglycaemic medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{insulin}{On insulin treatment - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{bpl}{Receiving at least one blood pressure lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{lld}{Receiving lipid lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{athromb}{Receiving antiplatelet or anticoagulant medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{...}{Set decimal place for integers. Default is 4. Optional.}
}
\value{
Returns either a single CVD risk estimate or a numeric vector of CVD risk estimates.
}
\description{
\code{PriorT2DRisk} calculates the 5 year risk of cardiovascular disease (CVD) (hospitalisation for acute coronary syndrome, heart failure, stroke or other cerebrovascular disease, peripheral vascular death, cardiovascular death),
for people with diabetes. This equation takes into account multiple diabetes-related variables. If a dataset of input values are not supplied, then individual values for each coefficient can be specified. If a dataset of input values are supplied, then a risk estimate is produced for each row of data, resulting in a numeric vector of the same length.
A specific format is required for each variable input value. Encoding may be required. See arguments.
}
\details{
When the parameter \code{dat} is supplied using a dataset, then parameters take variable names as input. For example, when a dataset is supplied, the parameter \code{age} requires the variable name \code{index_age} as input from the dataset.
When the parameter \code{dat} is not supplied, then parameters take actual values or labels as input. For example, when \code{dat} is not supplied, the parameter \code{age} requires a single numeric value between 30 and 79. This method calculates the 5-year risk estimate for a single individual.
The co-efficients for ethnicity apply only to the following labels (codes): "European" (1), "Maori" (2), "Pacific" (3), "Chinese" (42), "Indian" (43), "Fijian Indian" (43), and "Other Asian" (4). Individuals with ethnicity labels (or codes) that fall outside of these categories will not recieve a risk estimate.
To obtain a risk estimate please ensure that ethnicity is labelled (or encoded) as one of the above categories.
}
\examples{
# As vectoriser (Dataset Provided)
PriorT2DRisk(dat=DF, sex=sex, age=age, eth=eth, nzdep=nzdep, smoker=smoker, af=af, familyhx=familyhx, sbp=sbp, tchdl=tchdl,
          bmi=bmi, years=years, egfr=egfr, acr=acr, hba1c=hba1c, oral=oral, insulin=insulin, bpl=bpl, lld=lld, athromb=athromb)

# As calculator (Dataset not Provided)
PriorT2DRisk(sex=1, age=50, eth=2, nzdep=3, smoker=1, af=0, familyhx=1, sbp=123, tchdl=5, bmi=NA,
           years=5, egfr=80, acr=25, hba1c=40, oral=1, insulin=0, bpl=0, lld=0, athromb=1)

}
\seealso{
\code{\link{NoPriorCVDRisk}} Creates a 5 year CVD risk estimate for people without prior CVD using the published Lancet equation
\code{\link{NoPriorCVDRisk_BMI}} Creates a 5 year CVD risk estimate for people without prior CVD using the Ministry of Health's HISO equation containing BMI
\code{\link{PriorT2DRisk}} Creates a 5 year CVD risk estimate for people with prior Type-II diabetes using the Ministry of Health's HISO equation
\code{\link{MajorBleedRisk}} Creates a 5 year major bleeding risk estimate for people without prior CVD using the published AnnIntMed equation
\code{\link{PriorCVDRisk}} Creates a 5 year CVD risk estimate for people with prior CVD using the published Heart equation
\code{\link{PolicyCVDRisk}} Creates a 5 year CVD policy risk estimate for people in the general population using the publish IJE equation
}
