% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MajorBleedRisk.R
\name{MajorBleedRisk}
\alias{MajorBleedRisk}
\title{PREDICT Major Bleed (2018) Risk Score for People Without Prior CVD}
\usage{
MajorBleedRisk(dat, sex, age, eth, smoke, nzdep, af, familyhx, diabetes, sbp, tchdl, lld, bpl, cancer,
           gibleed, puddiag, alcohol, liver, puddrug, nsaid, steroids, ssri,...)
}
\arguments{
\item{dat}{A data.frame or data.table containing input data. Optional. See Details.}

\item{sex}{Sex or gender - input as labels M, Male, F, Female; or encode binary where 1 is male and 0 is female}

\item{age}{Age - input as numeric value between 30 and 74}

\item{eth}{Ethnicity - input as labels (or encode as) "European" (1), "Maori" (2), "Pacific" (3), "Chinese" (42), "Indian" (43), "Fijian Indian" (43), "Other South Asian" (43), or "Other Asian" (4).}

\item{nzdep}{Index of socioeconomic deprivation, specifically the New Zealand Deprivation Index - input as numeric quintile value between 1 (least deprived) and 5 (most deprived).}

\item{familyhx}{Family history of premature CVD - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{diabetes}{Diabetes status - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{sbp}{Systolic blood pressure - input as numeric value representing measured systolic blood pressure in mmHg}

\item{tchdl}{Total-HDL cholesterol ratio - input as numeric value representing most recent value of total:HDL cholesterol}

\item{lld}{Receiving lipid lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{bpl}{Receiving at least one blood pressure lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{cancer}{Any prior primary malignancy excluding squamous and basal cell skin cancers}

\item{gibleed}{Any prior gastrointestinal bleeding - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{puddiag}{Any prior non-bleeding and non-perforated peptic ulcer disease - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{alcohol}{Chronic high use of alcohol - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{liver}{Chronic liver disease or pancreatitis - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{puddrug}{Receiving at peptic ulcer disease medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{nsaid}{Receiving at least one nonsteroidal anti-inflammatary drug - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{steroids}{Receiving at least one corticosteroid medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{ssri}{Receiving at least one selective serotonine reuptake inhibitor - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{...}{Set decimal place for integers. Default is 4. Optional.}

\item{Smoking}{status - input as labels (or encode as) "Ex smoker" (2), "Ex-smoker" (2), "Ex" (2), "Current Smoker" (1), "Current" (1), "Smoker" (1), "Y" (1), "Yes" (1)}
}
\value{
Returns either a single bleeding risk estimate or a numeric vector of bleeding risk estimates.
}
\description{
\code{MajorBleedRisk} calculates the 5 year risk of major bleeding (gastrointestinal, intracranial, and other bleeds), for people without prior cardiovascular disease. This equation takes into account multiple diabetes-related variables. If a dataset of input values are not supplied, then individual values for each coefficient can be specified. If a dataset of input values are supplied, then a risk estimate is produced for each row of data, resulting in a numeric vector of the same length.
A specific format is required for each variable input value. Encoding may be required. See arguments.
}
\details{
When the parameter \code{dat} is supplied using a dataset, then parameters take variable names as input. For example, when a dataset is supplied, the parameter \code{age} requires the variable name \code{index_age} as input from the dataset.
When the parameter \code{dat} is not supplied, then parameters take actual values or labels as input. For example, when \code{dat} is not supplied, the parameter \code{age} requires a single numeric value between 30 and 79. This method calculates the 5-year risk estimate for a single individual.
The co-efficients for ethnicity apply only to the following labels (codes): "European" (1), "Maori" (2), "Pacific" (3), "Chinese" (42), "Indian" (43), "Fijian Indian" (43), and "Other Asian" (4). Individuals with ethnicity labels (or codes) that fall outside of these categories will not recieve a risk estimate.
To obtain a risk estimate please ensure that ethnicity is labelled (or encoded) as one of the above categories.
}
\examples{
# As calculator (dataset not provided)
MajorBleedRisk(sex=0, age=55, eth=21, smoker=1, nzdep=5, af=0, familyhx=1, diabetes=1, sbp=130, tchdl=5, lld=1,
            bpl=1, cancer=1, gibleed=1,puddiag=1, alcohol=0, liver=0, puddrug=0, nsaid=1, steroids=1, ssri=0)

# As vectoriser (dataset provided)
MajorBleedRisk(dat=DT, sex=sex, age=index_age, eth=eth_vars, smoker=smoking_status, nzdep=nzdep, af=af, familyhx=family_hx, diabetes=dm,
            sbp=index_sbp, tchdl=tchdl, lld=lld, bpl=bpl, cancer=hx_cancer, gibleed=hx_gibleed, puddiag=hx_hud, alcohol=alc, liver=hx_liver,
            puddrug=pudmx, nsaid=nsaid, steroids=steroidmx, ssri=ssri)

}
\seealso{
\code{\link{NoPriorCVDRisk}} Creates a 5 year CVD risk estimate for people without prior CVD using the published Lancet equation
\code{\link{NoPriorCVDRisk_BMI}} Creates a 5 year CVD risk estimate for people without prior CVD using the Ministry of Health's HISO equation containing BMI
\code{\link{PriorT2DRisk}} Creates a 5 year CVD risk estimate for people with prior Type-II diabetes using the Ministry of Health's HISO equation
\code{\link{MajorBleedRisk}} Creates a 5 year major bleeding risk estimate for people without prior CVD using the published AnnIntMed equation
\code{\link{PriorCVDRisk}} Creates a 5 year CVD risk estimate for people with prior CVD using the published Heart equation
\code{\link{PolicyCVDRisk}} Creates a 5 year CVD policy risk estimate for people in the general population using the publish IJE equation
}
