% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoPriorCVDRisk.R
\name{NoPriorCVDRisk}
\alias{NoPriorCVDRisk}
\title{PREDICT CVD (2017) Risk Score for People Without CVD}
\usage{
NoPriorCVDRisk(dat, sex, age, eth, nzdep, exsmoker, smoker, diabetes,
               af, familyhx, sbp, tchdl, bpl, lld, athrombi, ...)
}
\arguments{
\item{dat}{an optional data.frame or data.table containing input data (see details)}

\item{sex}{binary sex or gender}

\item{age}{age in years  (see details)}

\item{eth}{ethnicity (see details)}

\item{nzdep}{socio-economic deprivation (see details)}

\item{exsmoker}{ex-smoker or recently quit}

\item{smoker}{currently smoking}

\item{diabetes}{diabetes status}

\item{af}{atrial fibrillation status}

\item{familyhx}{family history of premature CVD}

\item{sbp}{measured systolic blood pressure in mmHg}

\item{tchdl}{most recent value of total:HDL cholesterol}

\item{bpl}{receiving at least one blood pressure lowering medication}

\item{lld}{receiving lipid lowering medication}

\item{athrombi}{receiving antiplatelet or anticoagulant medication}

\item{...}{further arguments (see values)}
}
\value{
returns either a single 5-year CVD risk estimate, or a numeric vector of risk estimates if \code{dat} is provided.
Input values for each parameter must conform to the following convention:

\item{sex}{label or encode as one of the following:
\itemize{
\item M, Male, 1
\item F, Female, 0
}}
\item{age}{numeric value for years of age between 20 and 110}
\item{eth}{label or encode as one of the following:
\itemize{
\item NZ European, European, NZEO, Euro, E, 1, 10, 11, or 12
\item Maori, NZMaori, NZ Maori, M, 2, or 21
\item Pacific, Pacific Islander, PI, P, 3, 30, 31, 32, 33, 34, 35, 36, or 37
\item Indian, Fijian Indian, South Asian, IN, I, or 43
\item Asian, Other Asian, SE Asian, East Asian, Chinese, ASN, A, 4, 40, 41, 42, or 44
\item note: Other Asian includes non-Indian South Asian
}}
\item{nzdep}{numeric value between 1 and 5}
\item{exsmoker}{label or encode as one of the following:
\itemize{
\item Y, Yes, Ex, Ex-smoker, Exsmoker, E, 1, T, TRUE
\item N, No, Non-smoker, Non, 0, F, FALSE
}}
\item{smoker}{label or encode as one of the following:
\itemize{
\item Y, Yes, Smoker, Current, S, 1, T, TRUE
\item N, No, Non-smoker, Non, 0, F, FALSE
}}
\item{diabetes, \cr af, familyhx}{label or encode as one of the following:
\itemize{
\item Y, Yes, 1, T, TRUE
\item N, No, 0, F, FALSE
}}
\item{bpl, lld,\cr athrombi}{label or encode as one of the following:
\itemize{
\item Y, Yes, 1, T, TRUE
\item N, No, 0, F, FALSE
}}
\item{sbp, tchdl}{numeric value of measured result. Note:
\itemize{
\item SBP and total:HDL values must be available
}}
\item{...}{further arguments:
\itemize{
\item \code{dp} numeric value to set decimal place; default is 4
\item \code{allow.age} logical. Whether or not age range is extended outside of 30 - 74; default is TRUE. If set to FALSE, then \code{NA} is returned as risk estimate.
\item \code{allow.na} logical. Whether or not missing values for binary variables and smoking status are treated as 0; default is TRUE. If set to FALSE, then \code{NA} is returned as risk estimate.
}}
}
\description{
\code{NoPriorCVDRisk} calculates the 5 year absolute risk of cardiovascular disease (CVD) for people without a history of atherosclerotic CVD.
In this version, BMI is not included as a predictor. The outcome of interest is the 5-year risk of a non-fatal or fatal CVD event, including hospitalisation
for coronary heart disease, stroke or other cerebrovascular disease (including transient ischaemic attack), peripheral vascular disease and heart failure,
or cardiovascular death.
}
\details{
When a dataset is supplied, a risk score is produced for each row of data, resulting in a numeric vector of the same length.
Each argument requires the variable name from the dataset \code{dat} that corresponds with the parameter. If the parameter \code{dat} is not supplied, then each argument is assigned an individual's
actual parameter value. \cr

The risk prediction equations were developed from a cohort of people aged 30 to 74 years. Additional analyses indicate that the sex-specific risk equations performed adequately in those aged 75-79 years.
People aged 18-29 years or 80 years and older, are outside the range used to derive the equation, and therefore risk will be even more of an approximation. The function
will calculate ages 18-29 as 30; and ages 80-110 as 79. All other age inputs are invalid and will return \code{NA}. \cr

The co-efficients for ethnicity apply only to the following groups: European, Maori, Pacific, Indian, and (non-Indian) Asian. To obtain a risk estimate, ensure that the
ethnicity input is either labelled or encoded using one of the values listed below (see values). All other inputs are invalid and will return \code{NA}. \cr

The scale for socioeconomic deprivation is derived from the New Zealand Index of Deprivation; a small area-based measure that combines census data relating to income,
home ownership, employment, qualifications, family structure, housing, access to transport and communications. The equations require NZDep to be categorised as quintiles,
with 1 being the least deprived and 5 being the most deprived.
}
\section{See Also}{

\code{\link{NoPriorCVDRisk}} \cr
\code{\link{NoPriorCVDRisk_BMI}} \cr
\code{\link{NoPriorCVDRisk_Policy}} \cr
\code{\link{NoPriorCVDBleedRisk}} \cr
\code{\link{NoPriorT2DRisk}} \cr
\code{\link{PostCVDRisk}} \cr
\code{\link{PostACSRisk}} \cr
}

\examples{
# As Calculator (dataset not provided)
NoPriorCVDRisk(sex="F", age=65, eth="Indian", smoker=0, nzdep=5,  diabetes=0,
               af=0, familyhx=1, lld=1, athrombi=1, bpl=1, sbp=118, tchdl=3.3)

NoPriorCVDRisk(sex=1, age=82, eth="Asian", exsmoker=NA, smoker=Y, nzdep=4, diabetes=N,
               af=1, familyhx=1, lld=T, athrombi=F, bpl=0, sbp=140, tchdl=3.3, dp = 5,
               allow.age = FALSE, allow.na = FALSE)

# As a vectoriser (dataset provided)
NoPriorCVDRisk(dat=TEST, sex=sex, age=age, eth=eth, nzdep=nzdep, exsmoker=smoker,
               smoker=smoker,  diabetes=diabetes, af=af, familyhx=familyhx, lld=lld,
               athrombi=athromb, bpl=bpl, sbp=sbp, tchdl=tchdl, allow.na =TRUE)

}
\references{
Pylypchuk R, Wells S, Kerr A, Poppe K, Riddell T, Harwood M, et al. Cardiovascular disease risk prediction equations in 400 000 primary care patients in New Zealand: a derivation and validation study. Lancet 2018;391:1897-907.

\href{https://www.thelancet.com/journals/lancet/article/PIIS0140-6736(18)30664-0/fulltext}{Full Article}
}
\author{
Billy Wu (R Developer) and Romana Pylypchuk (Principal Investigator)
}
