% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoPriorCVDRisk_Policy.R
\name{NoPriorCVDRisk_Policy}
\alias{NoPriorCVDRisk_Policy}
\title{VARIANZ CVD (2018) Policy-level Risk Score for People Without CVD}
\usage{
NoPriorCVDRisk_Policy(dat, sex, age, eth, nzdep, diabetes, af, bpl, lld, athromb,...)
}
\arguments{
\item{dat}{A data.frame or data.table containing input data. Optional. See Details.}

\item{sex}{Sex or gender - input as labels M, Male, F, Female; or encode binary where 1 is male and 0 is female}

\item{age}{Age - input as numeric value between 20 and 79. See age details if outside of this range.}

\item{eth}{Ethnicity - input as label or encoded value. See ethnicity details for all possible inputs.}

\item{nzdep}{Index of socioeconomic deprivation, specifically the New Zealand Deprivation Index - input as numeric quintile value between 1 (least deprived) and 5 (most deprived)}

\item{diabetes}{Diabetes status - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{af}{Atrial fibrillation status - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{bpl}{Receiving at least one blood pressure lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{lld}{Receiving lipid lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{athromb}{Receiving antiplatelet or anticoagulant medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{...}{Set decimal place for integers. Default is 4. Optional.}
}
\description{
\code{NoPriorCVDRisk_Policy} calculates the 5 year absolute risk of cardiovascular disease (CVD) for people without a history of atherosclerotic CVD that is intended
for use at policy or population level. The outcome of future CVD is defined as hospitalisation for acute coronary syndrome, heart failure, stroke or other
cerebrovascular disease, peripheral vascular disease, or cardiovascular death. If a dataset of input values are not supplied, then individual values for each
coefficient can be specified. If a dataset of input values are supplied, then a risk estimate is produced for each row of data, resulting in a numeric vector of the
same length. A specific format is required for each input value. Encoding may be required. See arguments.
}
\details{
When the parameter \code{dat} is supplied using a dataset, then parameters take variable names as input. For example, when a dataset is supplied, the parameter \code{age} requires the variable name \code{index_age} as input from the dataset.
When the parameter \code{dat} is not supplied, then parameters take actual values or labels as input. For example, when \code{dat} is not supplied, the parameter \code{age} requires a single numeric value between 30 and 79. This method calculates the 5-year risk estimate for a single individual.
}
\section{Age}{

The primary prevention risk prediction equations were developed from a cohort of people aged 30 to 74 years who were eligible for CVD risk prediction according to the 2003 CVD risk assessment and management guidelines and subsequent updates (New Zealand Guidelines Group 2003).
People aged 18-29 years and 80 years and older, the equation will only provide a very approximate estimate. However, a risk calculation may be potentially useful to guide clinical decision making.
As such, the equation will calculate ages 18-29 as 30; and ages 80-110 as 80.
People aged 75-79 years are also outside of the range for which the algorithms were developed. However, assessment of the equations performance (calibration) shows that they perform reasonably well.
Therefore, the equation will calculate ages 75-79 as per input.
}

\section{Ethnicity}{

The co-efficients for ethnicity apply only to the following groups: European, Maori, Pacific, Indian, and Other Individuals with ethnicity labels (or codes) that fall outside of these categories will not recieve a risk estimate.
To obtain a risk estimate, ensure that the ethnicity parameter is either labelled or encoded as one of the following:
\itemize{
\item NZ European, European, NZEO, Euro, E, 1, 10, 11, 12
\item Maori, NZMaori, NZ Maori, M, 2, 21
\item Pacific, Pacific Islander, PI, P, 3, 30, 31, 32, 33, 34, 35, 36, 37
\item Indian, Fijian Indian, South Asian, IN, I, 43
\item Other, Middle Eastern, African, Latin American, Latin, Asian, Other Asian, SE Asian, East Asian, Chinese, MELAA, ME, ASN,
      A, 4, 40, 41, 42, 44, 5, 51, 52, 53, 54, 61
}
}

\section{Value}{

Returns either a single CVD risk estimate or a numeric vector of CVD risk estimates.
}

\examples{
# As a calculator (dataset not provide)
NoPriorCVDRisk_Policy(sex="F", age=65, eth="Maori", nzdep=5, diabetes=0, af=0, bpl=1, lld=1, athromb=1, dp = 6)

# As Vectoriser (dataset provided)
NoPriorCVDRisk_Policy(dat=TEST, sex=sex, age=age, eth=eth, nzdep=nzdep, diabetes=diabetes, af=af, bpl=bpl, lld=lld,
                      athromb=athromb)

}
\references{
Mehta, S., Jackson, R., Pylypchuk, R., Poppe, K., Wells, S., & Kerr, A. J. (2018). Development and validation of alternative cardiovascular risk prediction equations for population health planning: a routine health data linkage study of 1.7 million New Zealanders. International journal of epidemiology, 47 (5), 1571-1584.

Full Article: \url{https://academic.oup.com/ije/article/47/5/1571/5053287}
}
\seealso{
\code{\link{NoPriorCVDRisk}} Creates a 5 year CVD risk estimate for people without prior CVD using the published Lancet equation.

\code{\link{NoPriorCVDRisk_BMI}} Creates a 5 year CVD risk estimate for people without prior CVD using the Ministry of Health's HISO equation containing BMI.

\code{\link{PriorT2DRisk}} Creates a 5 year CVD risk estimate for people with prior Type-II diabetes using the Ministry of Health's HISO equation.

\code{\link{MajorBleedRisk}} Creates a 5 year major bleeding risk estimate for people without prior CVD using the published AnnIntMed equation.

\code{\link{PriorCVDRisk}} Creates a 5 year CVD risk estimate for people with prior CVD using the published Heart equation.

\code{\link{PostACSRisk}} Creates a 5 year CVD risk estimate for people after an ACS event using the published Heart equation.
}
\author{
Billy Wu (R developer) and Suneela Mehta (Principal Investigator)
}
