% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoPriorCVDRisk_Policy.R
\name{NoPriorCVDRisk_Policy}
\alias{NoPriorCVDRisk_Policy}
\title{VARIANZ CVD (2018) Policy-level Risk Score for People Without CVD}
\usage{
NoPriorCVDRisk_Policy(dat, sex, age, eth, nzdep, diabetes, af,
                      bpl, lld, athrombi, ...)
}
\arguments{
\item{dat}{an optional data.frame or data.table containing input data (see details)}

\item{sex}{binary sex or gender}

\item{age}{age in years  (see details)}

\item{eth}{ethnicity (see details)}

\item{nzdep}{socio-economic deprivation (see details)}

\item{diabetes}{diabetes status}

\item{af}{atrial fibrillation status}

\item{bpl}{receiving at least one blood pressure lowering medication}

\item{lld}{receiving lipid lowering medication}

\item{athrombi}{receiving antiplatelet or anticoagulant medication}

\item{...}{further arguments (see values)}
}
\value{
returns either a single 5-year CVD risk estimate, or a numeric vector of risk estimates if \code{dat} is provided.
Input values for each parameter must conform to the following convention:

\item{sex}{label or encode as one of the following:
\itemize{
\item M, Male, 1
\item F, Female, 0
}}
\item{age}{numeric value for years of age between 20 and 110}
\item{eth}{label or encode as one of the following:
\itemize{
\item NZ European, European, NZEO, Euro, E, 1, 10, 11, or 12
\item Maori, NZMaori, NZ Maori, M, 2, or 21
\item Pacific, Pacific Islander, PI, P, 3, 30, 31, 32, 33, 34, 35, 36, or 37
\item Indian, Fijian Indian, IN, I, or 43
\item Other, Middle Eastern, African, Latin American, South American, Latin, Asian, Other Asian, SE Asian, East Asian,
Chinese, MELAA, ME, ASN, A, 4, 40, 41, 42, 44, 5, 51, 52, 53, 54, 61
\item note: Other Asian includes non-Indian South Asian
}}
\item{nzdep}{numeric value between 1 and 5}
\item{diabetes,\cr af}{label or encode as one of the following:
\itemize{
\item Y, Yes, 1, T, TRUE
\item N, No, 0, F, FALSE
}}
\item{bpl, lld,\cr athrombi}{label or encode as one of the following:
\itemize{
\item Y, Yes, 1, T, TRUE
\item N, No, 0, F, FALSE
}}
\item{...}{further arguments:
\itemize{
\item \code{dp} numeric value to set decimal place; default is 4
\item \code{allow.age} logical. Whether or not age range is extended outside of 30 - 74; default is TRUE. If set to FALSE, then \code{NA} is returned as risk estimate.
\item \code{allow.na} logical. Whether or not missing values for binary variables and smoking status are treated as 0; default is TRUE. If set to FALSE, then \code{NA} is returned as risk estimate.
}}
}
\description{
\code{NoPriorCVDRisk_Policy} calculates the 5 year absolute risk of cardiovascular disease (CVD) for people without a history of atherosclerotic CVD that
is intended for use at policy or general population level. The outcome of interest is the 5-year risk of a non-fatal or fatal CVD event, including hospitalisation
for coronary heart disease, stroke or other cerebrovascular disease (including transient ischaemic attack), peripheral vascular disease and heart failure,
or cardiovascular death.
}
\details{
\code{NoPriorCVDRisk_Policy} is intended to be used at the policy or general population level. As such, a dataset containing a population should be provided, and outputs should be summarised.
Using this function as a calculator for an individual is not recommended. When a dataset is supplied, a risk score is produced for each row of data, resulting in a numeric vector of the same length.
Each argument requires the variable name from the dataset \code{dat} that corresponds with the parameter. \cr

The risk prediction equations were developed from a cohort of people aged 30 to 74 years. Additional analyses indicate that the sex-specific risk equations performed adequately in those aged 75-79 years.
People aged 18-29 years or 80 years and older, are outside the range used to derive the equation, and therefore risk will be even more of an approximation. The function
will calculate ages 18-29 as 30; and ages 80-110 as 79. All other age inputs are invalid and will return \code{NA}. \cr

The co-efficients for ethnicity apply only to the following groups: European, Maori, Pacific, Indian, and (non-Indian) Asian. To obtain a risk estimate, ensure that the
ethnicity input is either labelled or encoded using one of the values listed below (see values). All other inputs are invalid and will return \code{NA}. \cr

The scale for socioeconomic deprivation is derived from the New Zealand Index of Deprivation; a small area-based measure that combines census data relating to income,
home ownership, employment, qualifications, family structure, housing, access to transport and communications. The equations require NZDep to be categorised as quintiles,
with 1 being the least deprived and 5 being the most deprived.
}
\section{See Also}{

\code{\link{NoPriorCVDRisk}} \cr
\code{\link{NoPriorCVDRisk_BMI}} \cr
\code{\link{NoPriorCVDRisk_Policy}} \cr
\code{\link{NoPriorCVDBleedRisk}} \cr
\code{\link{NoPriorT2DRisk}} \cr
\code{\link{PostCVDRisk}} \cr
\code{\link{PostACSRisk}} \cr
}

\examples{
# As a vectoriser (dataset provided)
NoPriorCVDRisk_Policy(dat=TEST, sex=sex, age=age, eth=eth, nzdep=nzdep, diabetes=diabetes, af=af, lld=lld,
                      athrombi=athromb, bpl=bpl, dp =6, allow.na = FALSE)

}
\references{
Mehta, S., Jackson, R., Pylypchuk, R., Poppe, K., Wells, S., & Kerr, A. J. (2018). Development and validation of alternative cardiovascular risk prediction equations for population health planning: a routine health data linkage study of 1.7 million New Zealanders. International journal of epidemiology, 47 (5), 1571-1584.

\href{https://academic.oup.com/ije/article/47/5/1571/5053287}{Full Article}
}
\author{
Billy Wu (R Developer) and Suneela Mehta (Principal Investigator)
}
