% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamCheck.R
\name{ParamCheck}
\alias{ParamCheck}
\title{General Parameter Checking for PredictRiskScores}
\usage{
ParamCheck(params, ignore.age)
}
\arguments{
\item{params}{character vector of expected parameters within this function}

\item{ignore.age}{logical; if \code{FALSE} then only valid ages of 30-74 are calculated}
}
\value{
\code{ParamCheck} returns warning messages if any invalid values are detected. In cases where input data is invalid, \code{ParamCheck} will
stop the function's operation and return an error message.
}
\description{
\code{ParamCheck} evaluates input parameters for names, values, and availability to ensure that a risk calculation can be carried out; else provide warnings, else stopped.
}
\details{
\code{ParamCheck} identifies the following errors and invalid inputs:
\enumerate{
\item missing arguments - operations stops
\item missing input data - operation stops
\item invalid sex, age, ethnicity, nzdep inputs
\item invalid binary inputs
\item invalid numeric inputs
\item invalid time inputs
}
}
\author{
Billy Wu (R Developer)
}
