% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostCVDRisk.R
\name{PostCVDRisk}
\alias{PostCVDRisk}
\title{PREDICT CVD (2019) Risk Score for People With CVD}
\usage{
PostCVDRisk(dat, sex, age, eth, nzdep, smoker, diabetes,
            af, hf, othervd, bpl, lld, athrombi, sbp, tchdl, bmi,
            scr, hba1c, cvddays...)
}
\arguments{
\item{dat}{an optional data.frame or data.table containing input data (see details)}

\item{sex}{binary sex or gender}

\item{age}{age in years  (see details)}

\item{eth}{ethnicity (see details)}

\item{nzdep}{socio-economic deprivation (see details)}

\item{smoker}{currently smoking}

\item{diabetes}{diabetes status}

\item{af}{atrial fibrillation status}

\item{hf}{heart failure status}

\item{othervd}{prior angina, peripheral vascular disease, or non-hospitalised cerebrovascular disease not associated with stroke or TIA}

\item{bpl}{receiving at least one blood pressure lowering medication}

\item{lld}{receiving lipid lowering medication}

\item{athrombi}{receiving antiplatelet or anticoagulant medication}

\item{sbp}{measured systolic blood pressure in mmHg}

\item{tchdl}{most recent value of total:HDL cholesterol}

\item{bmi}{body mass index in kg/m^2}

\item{scr}{most recent value of serum creatinine in micromol/L}

\item{hba1c}{most recent value of HbA1c in mmol/mol}

\item{cvddays}{time in days since most recent CVD event}

\item{...}{further arguments (see values)}
}
\value{
returns either a single 5-year CVD risk estimate, or a numeric vector of risk estimates if \code{dat} is provided.
Input values for each parameter must conform to the following convention:

\item{sex}{label or encode as one of the following:
\itemize{
\item M, Male, 1
\item F, Female, 0
}}
\item{age}{numeric value for years of age between 20 and 110}
\item{eth}{label or encode as one of the following:
\itemize{
\item NZ European, European, NZEO, Euro, E, 1, 10, 11, or 12
\item Maori, NZMaori, NZ Maori, M, 2, or 21
\item Pacific, Pacific Islander, PI, P, 3, 30, 31, 32, 33, 34, 35, 36, or 37
\item Indian, Fijian Indian, South Asian, IN, I, or 43
\item Asian, Other Asian, SE Asian, East Asian, Chinese, ASN, A, 4, 40, 41, 42, or 44
\item note: Other Asian includes non-Indian South Asian
}}
\item{nzdep}{numeric value between 1 and 5}
\item{smoker}{label or encode as one of the following:
\itemize{
\item Y, Yes, Smoker, 1, T, TRUE
\item N, No, Non-smoker, 0, F, FALSE
}}
\item{diabetes\cr af hf othervd bpl lld\cr athrombi}{label or encode as one of the following:
\itemize{
\item Y, Yes, 1, T, TRUE
\item N, No, 0, F, FALSE
}}
\item{sbp tchdl}{numeric value of measured result. Note:
\itemize{
\item SBP and total:HDL values must be available
}}
\item{bmi scr\cr hba1c}{numeric value of calculated BMI, and measured serum creatinine and hba1c. If a value is unknown, then input as \code{NA}}
\item{cvddays}{numeric value of number of days since last ACS event. Note:
\itemize{
\item If the days since most recent CVD event is not known, then keep as \code{NA}
\item Ensure that \code{othervd} is checked if any prior angina or peripheral vascular disease.
\item If \code{othervd} is selected, of if \code{cvddays} is unknown, then will default to 1826.
}}
\item{...}{further arguments:
\itemize{
\item \code{dp} numeric value to set decimal place; default is 4
\item \code{allow.age} logical. Whether or not age range is extended outside of 30 - 74; default is TRUE. If set to FALSE, then \code{NA} is returned as risk estimate.
\item \code{allow.na} logical. Whether or not missing values for binary variables and smoking status are treated as 0; default is TRUE. If set to FALSE, then \code{NA} is returned as risk estimate.
}}
}
\description{
\code{PostCVDRisk} calculates the 5 year absolute risk of cardiovascular disease (CVD) for people with a history of atherosclerotic CVD. The outcome of future CVD
is defined as hospitalisation for acute coronary syndrome, heart failure, stroke or other cerebrovascular disease, peripheral vascular disease, or cardiovascular
death.
}
\details{
When a dataset is supplied, a risk score is produced for each row of data, resulting in a numeric vector of the same length.
Each argument requires the variable name from the dataset \code{dat} that corresponds with the parameter. If the parameter \code{dat} is not supplied, then each argument is assigned an individual's
actual parameter value. \cr

The risk prediction equations were developed from a cohort of people aged 30 to 74 years. Additional analyses indicate that the sex-specific risk equations performed adequately in those aged 75-79 years.
People aged 18-29 years or 80 years and older, are outside the range used to derive the equation, and therefore risk will be even more of an approximation. The function
will calculate ages 18-29 as 30; and ages 80-110 as 79. All other age inputs are invalid and will return \code{NA}. \cr

The co-efficients for ethnicity apply only to the following groups: European, Maori, Pacific, Indian, and (non-Indian) Asian. To obtain a risk estimate, ensure that the
ethnicity input is either labelled or encoded using one of the values listed below (see values). All other inputs are invalid and will return \code{NA}. \cr

The scale for socioeconomic deprivation is derived from the New Zealand Index of Deprivation; a small area-based measure that combines census data relating to income,
home ownership, employment, qualifications, family structure, housing, access to transport and communications. The equations require NZDep to be categorised as quintiles,
with 1 being the least deprived and 5 being the most deprived.
}
\section{See Also}{

\code{\link{NoPriorCVDRisk}} \cr
\code{\link{NoPriorCVDRisk_BMI}} \cr
\code{\link{NoPriorCVDRisk_Policy}} \cr
\code{\link{NoPriorCVDBleedRisk}} \cr
\code{\link{NoPriorT2DRisk}} \cr
\code{\link{PostCVDRisk}} \cr
\code{\link{PostACSRisk}} \cr
}

\examples{
# As a calculator (dataset not provided)
PostCVDRisk(sex=0, age=65, eth=In, nzdep=5, smoker=0, diabetes=0, af=0,
            hf=1, othervd=0,  bpl=1, lld=1, athrombi=1, sbp=118, tchdl=3.3,
            bmi=NA, scr=52, hba1c=NA, cvddays=60)

PostCVDRisk(sex="F", age=76, eth="Indian", nzdep=5, smoker=0, diabetes=0,
            af=0, hf=1, othervd=0,  bpl=1, lld=1, athrombi=1, sbp=118,
            tchdl=3.3, bmi=NA, scr=52, hba1c=NA, cvddays=365,
            allow.age=F, allow.na=F)

# As a vectoriser (dataset provided)
PostCVDRisk(TEST, sex=sex, age=age, eth=eth, nzdep=nzdep, smoker=smoker,
            diabetes=diabetes, af=af, hf=hf, othervd=othervd, cvddays=days,
            bmi=bmi, sbp=sbp, tchdl=tchdl, hba1c=hba1c, scr=scr, bpl=bpl,
            lld=lld, athrombi=athromb)

}
\references{
This equation is yet to be published. However, it is an update to the previously published equation in Heart (Poppe et al. 2017).\cr
\href{https://heart.bmj.com/content/103/12/891.1}{Full Article}
}
\author{
Billy Wu (R Developer) and Katrina Poppe (Principal Investigator)
}
