% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorCVDRisk.R
\name{PriorCVDRisk}
\alias{PriorCVDRisk}
\title{PREDICT CVD risk Score for People With Prior CVD}
\usage{
PriorCVDRisk(dat, sex, age, eth, nzdep, smoker, diabetes, af, hf,
             days, bmi, sbp, tchdl, hba1c, scr, bpl, lld, athromb,...)
}
\arguments{
\item{dat}{A data.frame or data.table containing input data. Optional. See Details.}

\item{sex}{Sex or gender - input as labels M, Male, F, Female; or encode binary where 1 is male and 0 is female}

\item{age}{Age - input as numeric value between 35 and 79}

\item{eth}{Ethnicity - input as labels "Chinese", "Indian", "Other Asian", "Fijian Indian", "Maori", "Pacific", "Other", or "Unknown"}

\item{nzdep}{NZ deprivation index - input as numeric quintile value between 1 (least deprived) and 5 (most deprived)}

\item{diabetes}{Diabetes status - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{af}{Atrrial fibrillation status - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{hf}{Heart failure history - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{days}{Time since last CVD event - input as numeric value representing days}

\item{bmi}{Body mass index - input as numeric value representing BMI in KG/m^2}

\item{sbp}{Systolic blood pressure - input as numeric value representing actual systolic blood pressure}

\item{tchdl}{Total-HDL cholesterol ratio - input as numeric value representing actual lab total:HDL value}

\item{hba1c}{HbA1C - input as numeric value representing actual lab HbA1c value}

\item{scr}{Serum creatinine - input as numeric value representing actual lab serum creatinine value}

\item{bpl}{On blood pressure lowering treatment - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{lld}{On lipid lowering treatment - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{athromb}{On antithrombotic including antiplatelet or anticoagulant treatment - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{...}{Set decimal place for integers. Default is 4. Optional.}

\item{smoking}{Smoking status - input as labels "Y", "Yes", "Smoker"}
}
\value{
Returns either a single risk score or a numeric vector of risk scores.
}
\description{
\code{PriorCVDRisk} calculates the 5 year risk of CVD for people with a history of CVD. If a dataset of input values are not supplied, then individual values for each coefficent can be specified.
If a dataset of input values are supplied, then a score is produced for each row of data, resulting in a numeric vector of the same row length.
A specific format is required for each variable input value. Encoding may be required. See arguments.
}
\details{
When the parameter \code{dat} is supplied using a dataset, then parameters take variable names as input. For example, when a dataset is supplied, the parameter \code{age} requires the variable name \code{index_age} as input from the dataset.
When the parameter \code{dat} is not supplied and empty, then parameters take actual values or labels as input. For example, when there is no data, the parameter \code{age} requires a single numeric value between 35 and 79.This method calculates the risk score for a single individual.
}
\examples{
# As a calculator
PriorCVDRisk(sex="F", age=65, eth="Indian", nzdep=5, smoker=0, diabetes=0, af=0, hf=1, days=65,
             bmi=NA, sbp=118, tchdl=3.3, hba1c=NA, scr=52, bpl=1, lld=1, athromb=1)

#As Vectoriser (i.e. dataset provided)
PriorCVDRisk(DT, sex=view_ag_sex, age=index_age, eth=view_ag_eth, nzdep=index_en_nzdep_quintiles, smoker=pt_smoking, diabetes=imp_hx_diabetes,
             af=imp_hx_af, hf=imp_hx_heart_failure, days=days_since_event_predict, bmi=pt_en_bmi, sbp=sbp, tchdl=imp_index_tchdl_ratio,
             hba1c=hba1c_index2yr, scr=creatinine_index2yr, bpl=ph_all_bplds_prior_6mths, lld=ph_all_llds_prior_6mths, athromb=antithrombotics, dp = 6)
}
\seealso{
\code{\link{NoPriorCVDRisk}} can be used for people without a history of CVD.
}
