% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorT2DRisk.R
\name{PriorT2DRisk}
\alias{PriorT2DRisk}
\title{PREDICT CVD Type-II Diabetes (2018.1) Risk Score for People Without Prior CVD}
\usage{
PriorT2DRisk(dat, sex, age, eth, nzdep, smoker, af, familyhx,
             sbp, tchdl, bmi, years, egfr, acr, hba1c, oral, insulin,
             bpl, lld, athromb,...)
}
\arguments{
\item{dat}{A data.frame or data.table containing input data. Optional. See Details.}

\item{sex}{Sex or gender - input as labels M, Male, F, Female; or encode binary where 1 is male and 0 is female}

\item{age}{Age - input as numeric value between 30 and 74. See age details if outside of this range.}

\item{eth}{Ethnicity - input as label or encoded value. See ethnicity details for all possible inputs.}

\item{nzdep}{Index of socioeconomic deprivation, specifically the New Zealand Deprivation Index - input as numeric quintile value between 1 (least deprived) and 5 (most deprived).}

\item{smoker}{Current smoker - input as labels "Y", "Yes", "Smoker", or encode binary where 1 is "Yes"}

\item{af}{Atrial fibrillation status - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{familyhx}{Family history of premature CVD - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{sbp}{Systolic blood pressure - input as numeric value representing measured systolic blood pressure in mmHg}

\item{tchdl}{Total-HDL cholesterol ratio - input as numeric value representing most recent value of total:HDL cholesterol}

\item{bmi}{Body mass index - input as numeric value representing BMI in kg/m^2}

\item{years}{Years since diagnosis of type 2 diabetes - input as numeric value representing years}

\item{egfr}{Calculated or measured estimated glomerular filtration rate - input as numeric value representing most recent value eGFR value in mL/min/1.73m2}

\item{acr}{Calculated or Measured albumin to creatinine ratio - input as numeric value representing most recent ACR value in mg/mmol}

\item{hba1c}{Haemoglobin A1C - input as numeric value representing most recent value of HbA1c in mmol/mol}

\item{oral}{On oral hypoglycaemic medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{insulin}{On insulin treatment - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{bpl}{Receiving at least one blood pressure lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{lld}{Receiving lipid lowering medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{athromb}{Receiving antiplatelet or anticoagulant medication - input as label "Y", "Yes", or encode binary where 1 is "Yes"}

\item{...}{Set decimal place for integers. Default is 4. Optional.}
}
\description{
\code{PriorT2DRisk} calculates the 5 year risk of cardiovascular disease (CVD) (hospitalisation for acute coronary syndrome, heart failure, stroke or other cerebrovascular disease, peripheral vascular death, cardiovascular death),
for people with diabetes. This equation takes into account multiple diabetes-related variables. If a dataset of input values are not supplied, then individual values for each coefficient can be specified. If a dataset of input values are supplied, then a risk estimate is produced for each row of data, resulting in a numeric vector of the same length.
A specific format is required for each variable input value. Encoding may be required. See arguments.
}
\details{
When the parameter \code{dat} is supplied using a dataset, then parameters take variable names as input. For example, when a dataset is supplied, the parameter \code{age} requires the variable name \code{index_age} as input from the dataset.
When the parameter \code{dat} is not supplied, then parameters take actual values or labels as input. For example, when \code{dat} is not supplied, the parameter \code{age} requires a single numeric value between 30 and 79. This method calculates the 5-year risk estimate for a single individual.
}
\section{Age}{

The primary prevention risk prediction equations were developed from a cohort of people aged 30 to 74 years who were eligible for CVD risk prediction according to the 2003 CVD risk assessment and management guidelines and subsequent updates (New Zealand Guidelines Group 2003).
People aged 18-29 years and 80 years and older, the equation will only provide a very approximate estimate. However, a risk calculation may be potentially useful to guide clinical decision making.
As such, the equation will calculate ages 18-29 as 30; and ages 80-110 as 80.
People aged 75-79 years are also outside of the range for which the algorithms were developed. However, assessment of the equations performance (calibration) shows that they perform reasonably well.
Therefore, the equation will calculate ages 75-79 as per input.
}

\section{Ethnicity}{

The co-efficients for ethnicity apply only to the following groups: European, Maori, Pacific, Indian, and Asian. Individuals with ethnicity labels (or codes) that fall outside of these categories will not recieve a risk estimate.
To obtain a risk estimate, ensure that the ethnicity parameter is either labelled (not case-sensitive) or encoded as one of the following:
\itemize{
\item NZ European, European, NZEO, Euro, E, 1, 10, 11, 12
\item Maori, NZMaori, NZ Maori, M, 2, 21
\item Pacific, Pacific Islander, PI, P, 3, 30, 31, 32, 33, 34, 35, 36, 37
\item Indian, Fijian Indian, South Asian, IN, I, 43
\item Asian, Other Asian, SE Asian, East Asian, Chinese, ASN, A, 4, 40, 41, 42, 44
}
}

\section{Value}{

Returns either a single CVD risk estimate or a numeric vector of CVD risk estimates.
}

\examples{
# As calculator (Dataset not Provided)
PriorT2DRisk(sex="M", age=35, eth=2, nzdep=5, smoker=1, af=1, familyhx=1, sbp=120, tchdl=3.3,
             bmi=27, years=1, egfr=78, acr=1, hba1c=48, oral=0, insulin=0, bpl=0, lld=0,
             athromb=0)

# As vectoriser (Dataset Provided)
PriorT2DRisk(dat=DF, sex=sex, age=age, eth=eth, nzdep=nzdep, smoker=smoker, af=af, familyhx=familyhx,
             sbp=sbp, tchdl=tchdl, bmi=bmi, years=years, egfr=egfr, acr=acr, hba1c=hba1c, oral=oral,
             insulin=insulin, bpl=bpl, lld=lld, athromb=athromb)

}
\references{
New Zealand Ministry of Health: HISO 10071:2019 Cardiovascular Disease Risk Assessment Data Standard

\url{https://www.health.govt.nz/publication/hiso-100712019-cardiovascular-disease-risk-assessment-data-standard}
}
\seealso{
\code{\link{NoPriorCVDRisk}} Creates a 5 year CVD risk estimate for people without prior CVD using the published Lancet equation.

\code{\link{NoPriorCVDRisk_BMI}} Creates a 5 year CVD risk estimate for people without prior CVD using the Ministry of Health's HISO equation containing BMI.

\code{\link{MajorBleedRisk}} Creates a 5 year major bleeding risk estimate for people without prior CVD using the published AnnIntMed equation.

\code{\link{PriorCVDRisk}} Creates a 5 year CVD risk estimate for people with prior CVD using the published Heart equation.

\code{\link{PolicyCVDRisk}} Creates a 5 year CVD policy risk estimate for people in the general population using the publish IJE equation.

\code{\link{PostACSRisk}} Creates a 5 year CVD risk estimate for people after an ACS event using the published Heart equation.
}
\author{
Billy Wu (R developer) and Romana Pylypchuk (Principle Investigator)
}
