% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorT2DRisk.R
\name{PriorT2DRisk}
\alias{PriorT2DRisk}
\title{PREDICT CVD Type-II Diabetes (2018.1) Risk Score for People Without Prior CVD}
\usage{
PriorT2DRisk(sex, age, eth, nzdep, smoker, af, familyhx,
             lld, athrombi, bpl, oral, insulin, sbp, tchdl, bmi,
             egfr, acr, hba1c, years, ...)
}
\arguments{
\item{dat}{an optional data.frame or data.table containing input data (see details)}

\item{sex}{binary sex or gender}

\item{age}{age in years  (see details)}

\item{eth}{ethnicity (see details)}

\item{nzdep}{socio-economic deprivation (see details)}

\item{smoker}{currently smoking}

\item{af}{atrial fibrillation status}

\item{familyhx}{family history of premature CVD}

\item{lld}{receiving lipid lowering medication}

\item{athrombi}{receiving antiplatelet or anticoagulant medication}

\item{bpl}{receiving at least one blood pressure lowering medication}

\item{oral}{receiving oral hypoglycaemic medication}

\item{insulin}{receiving insulin treatment}

\item{sbp}{measured systolic blood pressure in mmHg}

\item{tchdl}{most recent value of total:HDL cholesterol}

\item{bmi}{body mass index in kg/m^2}

\item{egfr}{most recent calculated value of eGFRvalue in mL/min/1.73m2}

\item{acr}{most recent value of ACR value in mg/mmol}

\item{hba1c}{most recent value of HbA1c in mmol/mol}

\item{years}{years since diagnosis of type 2 diabetes}

\item{...}{further arguments (see values)}
}
\value{
returns either a single 5-year CVD risk estimate, or a numeric vector of risk estimates if \code{dat} is provided.
Input values for each parameter must conform to the following convention:

\item{sex}{label or encode as one of the following:
\itemize{
\item M, Male, 1
\item F, Female, 0
}}
\item{age}{numeric value for years of age between 20 and 110}
\item{eth}{label or encode as one of the following:
\itemize{
\item NZ European, European, NZEO, Euro, E, 1, 10, 11, or 12
\item Maori, NZMaori, NZ Maori, M, 2, or 21
\item Pacific, Pacific Islander, PI, P, 3, 30, 31, 32, 33, 34, 35, 36, or 37
\item Indian, Fijian Indian, South Asian, IN, I, or 43
\item Asian, Other Asian, SE Asian, East Asian, Chinese, ASN, A, 4, 40, 41, 42, or 44
\item note: Other Asian includes non-Indian South Asian
}}
\item{nzdep}{numeric value between 1 and 5}
\item{smoker}{label or encode as one of the following:
\itemize{
\item Y, Yes, Smoker, Current, S, 1, T, TRUE
\item N, No, Non-smoker, 0, F, FALSE
}}
\item{af, familyhx}{label or encode as one of the following:
\itemize{
\item Y, Yes, 1, T, TRUE
\item N, No, 0, F, FALSE
}}
\item{bpl, lld,\cr athrombi}{label or encode as one of the following:
\itemize{
\item Y, Yes, 1, T, TRUE
\item N, No, 0, F, FALSE
}}
\item{oral, insulin}{label or encode as one of the following:
\itemize{
\item Y, Yes, 1, T, TRUE
\item N, No, 0, F, FALSE
}}
\item{bmi, sbp, tchdl,\cr egfr, acr, hba1c}{numeric value of measured result. Note:
\itemize{
\item all values must be available
}}
\item{years}{numeric value of number of years since T2D diagnosis}
\item{...}{further arguments:
\itemize{
\item \code{dp} numeric value to set decimal place; default is 4
\item \code{allow.age} logical. Whether or not age range is extended outside of 30 - 74; default is TRUE. If set to FALSE, then \code{NA} is returned as risk estimate.
\item \code{allow.na} logical. Whether or not missing values for binary variables and smoking status are treated as 0; default is TRUE. If set to FALSE, then \code{NA} is returned as risk estimate.
}}
}
\description{
\code{PriorT2DRisk} calculates the 5 year risk of cardiovascular disease (CVD) (hospitalisation for acute coronary syndrome, heart failure, stroke or other cerebrovascular disease, peripheral vascular death, cardiovascular death),
for people with diabetes. This equation takes into account multiple diabetes-related variables. The outcome of interest is the 5-year risk of a non-fatal or fatal CVD event, including hospitalisation
for coronary heart disease, stroke or other cerebrovascular disease (including transient ischaemic attack), peripheral vascular disease and heart failure,
or cardiovascular death.
}
\details{
When a dataset is supplied, a risk score is produced for each row of data, resulting in a numeric vector of the same length.
Each argument requires the variable name from the dataset \code{dat} that corresponds with the parameter. If the parameter \code{dat} is not supplied, then each argument is assigned an individual's
actual parameter value. \cr

The risk prediction equations were developed from a cohort of people aged 30 to 74 years. Additional analyses indicate that the sex-specific risk equations performed adequately in those aged 75-79 years.
People aged 18-29 years or 80 years and older, are outside the range used to derive the equation, and therefore risk will be even more of an approximation. The function
will calculate ages 18-29 as 30; and ages 80-110 as 79. All other age inputs are invalid and will return \code{NA}. \cr

The co-efficients for ethnicity apply only to the following groups: European, Maori, Pacific, Indian, and (non-Indian) Asian. To obtain a risk estimate, ensure that the
ethnicity input is either labelled or encoded using one of the values listed below (see values). All other inputs are invalid and will return \code{NA}. \cr

The scale for socioeconomic deprivation is derived from the New Zealand Index of Deprivation; a small area-based measure that combines census data relating to income,
home ownership, employment, qualifications, family structure, housing, access to transport and communications. The equations require NZDep to be categorised as quintiles,
with 1 being the least deprived and 5 being the most deprived.
}
\section{See Also}{

\code{\link{NoPriorCVDRisk}} \cr
\code{\link{NoPriorCVDRisk_BMI}} \cr
\code{\link{NoPriorCVDRisk_Policy}} \cr
\code{\link{NoPriorCVDBleedRisk}} \cr
\code{\link{NoPriorT2DRisk}} \cr
\code{\link{PostCVDRisk}} \cr
\code{\link{PostACSRisk}} \cr
}

\examples{
# As calculator (Dataset not provided)
PriorT2DRisk(sex="M", age=35, eth=2, nzdep=5, smoker=1, af=1, familyhx=1, lld=0,
             athrombi=0, bpl=0, oral=0, insulin=0, sbp=120, tchdl=3.3, bmi=27,
             egfr=78, acr=1, hba1c=48, years=1)

PriorT2DRisk(sex=0, age=75, eth=PI, nzdep=3, smoker=0, af=F, familyhx=0, lld=Y,
             athrombi=0, bpl=F, oral=T, insulin=0, sbp=130, tchdl=4, bmi=31,
             egfr=92, acr=1.4, hba1c=56, years=3)

# As a vectoriser (Dataset provided)
PriorT2DRisk(dat=DF, sex=sex, age=age, eth=eth, nzdep=nzdep, smoker=smoker,
             af=af, familyhx=familyhx, sbp=sbp, tchdl=tchdl, bmi=bmi, years=years,
             egfr=egfr, acr=acr, hba1c=hba1c, oral=oral, insulin=insulin, bpl=bpl,
             lld=lld, athrombi=athrombi)

}
\references{
New Zealand Ministry of Health: HISO 10071:2019 Cardiovascular Disease Risk Assessment Data Standard

\href{https://www.health.govt.nz/publication/hiso-100712019-cardiovascular-disease-risk-assessment-data-standard}{HISO Document}
}
\author{
Billy Wu (R Developer) and Romana Pylypchuk (Principal Investigator)
}
